risk-engine/requirements.txt
# Core ML and Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0

# Data I/O
pyarrow>=10.0.0
fastparquet>=0.8.0

# Pipeline and Reproducibility
dvc>=2.30.0
pyyaml>=6.0

# Evaluation and Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Explainability
shap>=0.41.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
pre-commit>=2.20.0

# Optional: Serving and Monitoring
fastapi>=0.85.0
uvicorn>=0.18.0
streamlit>=1.15.0
mlflow>=2.0.0

# Utilities
joblib>=1.2.0
tqdm>=4.64.0
